+++ polars_hist_db/dataset/__init__.py
@@ -1,3 +1,3 @@
-from .workflow import run_datasets
+from .entrypoint import run_datasets
 
 __all__ = ["run_datasets"]
