+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 name = "polars-hist-db"
 version = "0.6.0"
-description = "dsv --> dataframe <--> bitemporal-tables"
+description = "(dsv,jetstream) --> dataframe <--> bitemporal-tables"
 authors = ["jr200 <jayshan+git@gmail.com>"]
 repository = "https://github.com/jr200/polars-hist-db"
 readme = "README.md"
